1、修改选择传感器对话框点击自动识别崩溃问题。 2、设置选择传感器对话框的显示部分。 3、屏蔽警告信息。

+++ Src/Edislab/Edislab Pro/GridDataManager.cpp
@@ -55,7 +55,7 @@ void CGridDataManager::SetGroupInfo(GROUPDATA GroupData)
 
 void CGridDataManager::SetColumnInfo(int nColIndex,COLUMNDATA ColData)
 {
-	int nCount = m_vecGridColumnInfo.size();
+	int nCount = static_cast<int>(m_vecGridColumnInfo.size());
 	if(nColIndex < 0||nColIndex >= nCount)
 		return;
